Set up for verification
- more setup for verification handling
