+++ src/pages/auth/verified.tsx
@@ -1,0 +1,9 @@
+
+import EmailVerifiedLayout from "@/layouts/auth/EmailVerifiedLayout";
+export default function EmailVerified() {
+    return (
+        <>
+        <EmailVerifiedLayout />
+        </>
+    );
+}
